upload & download lectures part 1, multer grid fs outdated

+++ backend/controllers/lectureController.js
@@ -1,5 +1,6 @@
 import mongoose from 'mongoose';
 import Lecture from '../models/Lecture.js';
+import { gfs } from '../config/gridfs.js';
 
 export const getLectures = async (req, res) => {
 
@@ -19,22 +20,38 @@ export const getLectures = async (req, res) => {
 };
 
 export const createLecture = async (req, res) => {
-  const { title, filePath, course } = req.body;
-
-  if(!title || !filePath || !course) {
-    return res.status(400).json({ success: false, message: "Please provide all fields" });
-  }
+  const { title, course } = req.body;
 
   try {
+
+    console.log(req.file);
+
+    if(!req.file) {
+      return res.status(400).json({
+        success: false,
+        message: "Please upload a file"
+      })
+    }
+
+    if(!title || !course) {
+      return res.status(400).json({ success: false, message: "Please provide all fields" });
+    }
     
     const newLecture = new Lecture({
-      title,
-      filePath,
-      course
+      title: title,
+      filePath: req.file.id,
+      course: course
     });
 
     await newLecture.save();
-    res.status(201).json({ success: true, data: newLecture });
+
+    res.status(201).json({
+      success: true,
+      data: {
+        ...newLecture._doc,
+        originalName: req.file.metadata.originalName
+      }
+    });
   } catch(error) {
     console.log("Error creating lecture: ", error.message);
     res.status(500).json({ success: false, message: "Server error"});
@@ -66,6 +83,13 @@ export const deleteLecture = async (req, res) => {
   }
 
   try {
+    const lecture = await Lecture.findById(id);
+    if(!lecture) {
+      return req.status(404).json({ success: false, message: "Lecture not found"});
+    }
+
+    await gfs.delete(new mongoose.Types.ObjectId(lecture.filePath));
+
     await Lecture.findByIdAndDelete(id);
     res.status(200).json({ success: true, message: "Deleted" });
   } catch (error) {
@@ -73,3 +97,32 @@ export const deleteLecture = async (req, res) => {
     res.status(500).json({ success: false, message: "Server error"});
   }
 }
+
+export const downloadLecture = async (req, res) => {
+  try {
+    const lecture = await Lecture.findById(req.params.id);
+    if(!lecture) {
+      return res.status(404).json({
+        success: false,
+        message: "Lecture not found"
+      });
+    }
+
+    const file = await gfs.find({ _id: new mongoose.Types.ObjectId(lecture.filePath) }).toArray();
+    if(!file || file.length === 0) {
+      return res.status(404).json({
+        success: false,
+        message: "File not found"
+      });
+    }
+
+    res.header('Content-Type', 'application/pdf');
+    res.header('Content-Disposition', `attachment; filename="${file[0].metadata.originalName}"`);
+
+    const downloadStream = gfs.openDownloadStream(new mongoose.Types.ObjectId(lecture.filePath));
+    downloadStream.pipe(res);
+  } catch (error) {
+    console.log("Error downloading file: ", error.message);
+    res.status(500).json({ success: false, message: "Server error"});
+  }
+};
